[#184] 복합 커서 기반 페이지네이션 처리

+++ src/main/java/com/team6/hrbank/service/ChangeLogServiceImpl.java
@@ -82,7 +82,7 @@ public class ChangeLogServiceImpl implements ChangeLogService {
         nextCursor,
         nextIdAfter,
         condition.size(),
-        contents.size(),
+        (long) contents.size(),
         hasNext
     );
   }
